Updated ibc paths
