+++ src/models/index.ts
@@ -37,32 +37,32 @@ import MsgDtagRefuseTransfer from './msg/profiles/msg_dtag_refuse_transfer';
 import MsgDtagTransferRequest from './msg/profiles/msg_dtag_transfer_request';
 import MsgSaveProfile from './msg/profiles/msg_save_profile';
 import MsgUnblockUser from './msg/profiles/msg_unblock_user';
-import MsgCreateClient from './msg/client/msg_create_client';
-import MsgUpdateClient from './msg/client/msg_update_client';
-import MsgUpgradeClient from './msg/client/msg_upgrade_client';
-import MsgSubmitMisbehaviour from './msg/client/msg_submit_misbehaviour';
-import MsgHeight from './msg/client/msg_height';
-import MsgAcknowledgement from './msg/channel/msg_acknowledgement';
-import MsgChannelCloseConfirm from './msg/channel/msg_channel_close_confirm';
-import MsgChannelCloseInit from './msg/channel/msg_channel_close_init';
-import MsgChannelOpenAck from './msg/channel/msg_channel_open_ack';
-import MsgChannelOpenConfirm from './msg/channel/msg_channel_open_confirm';
-import MsgChannelOpenInit from './msg/channel/msg_channel_open_init';
-import MsgChannelOpenTry from './msg/channel/msg_channel_open_try';
-import MsgChannel from './msg/channel/msg_channel';
-import MsgCounterpartyChannel from './msg/channel/msg_counterparty';
-import MsgPacket from './msg/channel/msg_packet';
-import MsgReceivePacket from './msg/channel/msg_receive_packet';
-import MsgTimeout from './msg/channel/msg_timeout';
-import MsgTimeoutOnClose from './msg/channel/msg_timeout_on_close';
-import MsgConnectionEnd from './msg/connection/msg_connection_end';
-import MsgConnectionOpenAck from './msg/connection/msg_connection_open_ack';
-import MsgConnectionOpenConfirm from './msg/connection/msg_connection_open_confirm';
-import MsgConnectionOpenInit from './msg/connection/msg_connection_open_init';
-import MsgConnectionOpenTry from './msg/connection/msg_connection_open_try';
-import MsgCounterpartyConnection from './msg/connection/msg_counterparty';
-import MsgVersion from './msg/connection/msg_version';
-import MsgTransfer from './msg/transfer/msg_transfer';
+import MsgCreateClient from './msg/ibc/msg_client_create_client';
+import MsgUpdateClient from './msg/ibc/msg_client_update_client';
+import MsgUpgradeClient from './msg/ibc/msg_client_upgrade_client';
+import MsgSubmitMisbehaviour from './msg/ibc/msg_client_submit_misbehaviour';
+import MsgHeight from './msg/ibc/msg_client_height';
+import MsgAcknowledgement from './msg/ibc/msg_channel_acknowledgement';
+import MsgChannelCloseConfirm from './msg/ibc/msg_channel_close_confirm';
+import MsgChannelCloseInit from './msg/ibc/msg_channel_close_init';
+import MsgChannelOpenAck from './msg/ibc/msg_channel_open_ack';
+import MsgChannelOpenConfirm from './msg/ibc/msg_channel_open_confirm';
+import MsgChannelOpenInit from './msg/ibc/msg_channel_open_init';
+import MsgChannelOpenTry from './msg/ibc/msg_channel_open_try';
+import MsgChannel from './msg/ibc/msg_channel';
+import MsgCounterpartyChannel from './msg/ibc/msg_channel_counterparty';
+import MsgPacket from './msg/ibc/msg_channel_packet';
+import MsgReceivePacket from './msg/ibc/msg_channel_receive_packet';
+import MsgTimeout from './msg/ibc/msg_channel_timeout';
+import MsgTimeoutOnClose from './msg/ibc/msg_channel_timeout_on_close';
+import MsgConnectionEnd from './msg/ibc/msg_connection_end';
+import MsgConnectionOpenAck from './msg/ibc/msg_connection_open_ack';
+import MsgConnectionOpenConfirm from './msg/ibc/msg_connection_open_confirm';
+import MsgConnectionOpenInit from './msg/ibc/msg_connection_open_init';
+import MsgConnectionOpenTry from './msg/ibc/msg_connection_open_try';
+import MsgCounterpartyConnection from './msg/ibc/msg_connection_counterparty';
+import MsgVersion from './msg/ibc/msg_connection_version';
+import MsgTransfer from './msg/ibc_transfer/msg_transfer';
 
 export {
   BigDipperNetwork,
