+++ src/models/msg/ibc/msg_channel_acknowledgement.ts
@@ -1,31 +1,28 @@
 import { Categories } from '../types';
 
-class MsgReceivePacket {
+class MsgAcknowledgement {
     public category: Categories;
     public type: string;
     public signer: string;
     public sourceChannel: string;
-    public destinationChannel: string;
     public json: any;
 
     constructor(payload: any) {
-      this.category = 'channel';
+      this.category = 'ibc';
       this.type = payload.type;
       this.signer = payload.signer;
       this.sourceChannel = payload.sourceChannel;
-      this.destinationChannel = payload.destinationChannel;
       this.json = payload.json;
     }
 
     static fromJson(json: any) {
-      return new MsgReceivePacket({
+      return new MsgAcknowledgement({
         json,
         type: json['@type'],
         signer: json.signer,
         sourceChannel: json.packet?.source_channel,
-        destinationChannel: json.packet?.destination_channel,
       });
     }
 }
 
-export default MsgReceivePacket;
+export default MsgAcknowledgement;
