+++ src/models/msg/ibc/msg_channel_close_confirm.ts
@@ -1,31 +1,31 @@
 import { Categories } from '../types';
 
-class MsgConnectionOpenInit {
+class MsgChannelCloseConfirm {
     public category: Categories;
     public type: string;
     public signer: string;
-    public clientId: string;
-    public counterparty: string;
+    public channelId: string;
+    public portId: string;
     public json: any;
 
     constructor(payload: any) {
-      this.category = 'connection';
+      this.category = 'ibc';
       this.type = payload.type;
       this.signer = payload.signer;
-      this.clientId = payload.clientId;
-      this.counterparty = payload.counterparty;
+      this.channelId = payload.channelId;
+      this.portId = payload.portId;
       this.json = payload.json;
     }
 
     static fromJson(json: any) {
-      return new MsgConnectionOpenInit({
+      return new MsgChannelCloseConfirm({
         json,
         type: json['@type'],
         signer: json.signer,
-        clientId: json.client_id,
-        counterparty: json.counterparty,
+        channelId: json.channel_id,
+        portId: json.port_id,
       });
     }
 }
 
-export default MsgConnectionOpenInit;
+export default MsgChannelCloseConfirm;
