+++ src/models/msg/ibc/msg_channel_close_init.ts
@@ -1,28 +1,31 @@
 import { Categories } from '../types';
 
-class MsgConnectionOpenConfirm {
+class MsgChannelCloseInit {
     public category: Categories;
     public type: string;
     public signer: string;
-    public connectionId: string;
+    public channelId: string;
+    public portId: string;
     public json: any;
 
     constructor(payload: any) {
-      this.category = 'connection';
+      this.category = 'ibc';
       this.type = payload.type;
       this.signer = payload.signer;
-      this.connectionId = payload.connectionId;
+      this.channelId = payload.channelId;
+      this.portId = payload.portId;
       this.json = payload.json;
     }
 
     static fromJson(json: any) {
-      return new MsgConnectionOpenConfirm({
+      return new MsgChannelCloseInit({
         json,
         type: json['@type'],
         signer: json.signer,
-        connectionId: json.connection_id,
+        channelId: json.channel_id,
+        portId: json.port_id,
       });
     }
 }
 
-export default MsgConnectionOpenConfirm;
+export default MsgChannelCloseInit;
