+++ src/models/msg/ibc/msg_channel_counterparty.ts
@@ -1,31 +1,25 @@
 import { Categories } from '../types';
 
-class MsgUpdateClient {
+class MsgCounterpartyChannel {
     public category: Categories;
     public type: string;
     public signer: string;
-    public chainId: string;
-    public clientId: string;
     public json: any;
 
     constructor(payload: any) {
-      this.category = 'client';
+      this.category = 'ibc';
       this.type = payload.type;
       this.signer = payload.signer;
-      this.chainId = payload.chainId;
-      this.clientId = payload.clientId;
       this.json = payload.json;
     }
 
     static fromJson(json: any) {
-      return new MsgUpdateClient({
+      return new MsgCounterpartyChannel({
         json,
         type: json['@type'],
         signer: json.signer,
-        chainId: json.header?.signed_header?.header?.chain_id,
-        clientId: json.client_id,
       });
     }
 }
 
-export default MsgUpdateClient;
+export default MsgCounterpartyChannel;
