+++ src/models/msg/ibc/msg_channel_open_try.ts
@@ -1,37 +1,34 @@
 import { Categories } from '../types';
 
-class MsgChannelCloseInit {
+class MsgChannelOpenTry {
     public category: Categories;
     public type: string;
     public signer: string;
-    public channelId: string;
+    public channel: string;
     public portId: string;
-    public counterpartyChannelId: string;
     public counterpartyVersion: string;
     public json: any;
 
     constructor(payload: any) {
-      this.category = 'channel';
+      this.category = 'ibc';
       this.type = payload.type;
       this.signer = payload.signer;
-      this.channelId = payload.channelId;
+      this.channel = payload.channel;
       this.portId = payload.portId;
-      this.counterpartyChannelId = payload.counterpartyChannelId;
       this.counterpartyVersion = payload.counterpartyVersion;
       this.json = payload.json;
     }
 
     static fromJson(json: any) {
-      return new MsgChannelCloseInit({
+      return new MsgChannelOpenTry({
         json,
         type: json['@type'],
         signer: json.signer,
-        channelId: json.channel_id,
+        channel: json.channel,
         portId: json.port_id,
-        counterpartyChannelId: json.counterparty_channel_id,
         counterpartyVersion: json.counterparty_version,
       });
     }
 }
 
-export default MsgChannelCloseInit;
+export default MsgChannelOpenTry;
