+++ src/models/msg/ibc/msg_client_height.ts
@@ -1,28 +1,25 @@
 import { Categories } from '../types';
 
-class MsgSubmitMisbehaviour {
+class MsgHeight {
     public category: Categories;
     public type: string;
     public signer: string;
-    public clientId: string;
     public json: any;
 
     constructor(payload: any) {
-      this.category = 'client';
+      this.category = 'ibc';
       this.type = payload.type;
       this.signer = payload.signer;
-      this.clientId = payload.clientId;
       this.json = payload.json;
     }
 
     static fromJson(json: any) {
-      return new MsgSubmitMisbehaviour({
+      return new MsgHeight({
         json,
         type: json['@type'],
         signer: json.signer,
-        clientId: json.client_id,
       });
     }
 }
 
-export default MsgSubmitMisbehaviour;
+export default MsgHeight;
