+++ src/screens/transaction_details/components/index.ts
@@ -26,32 +26,32 @@ import DtagCancelTransfer from './msg/profiles/dtag_cancel_transfer';
 import DtagRefuseTransfer from './msg/profiles/dtag_refuse_transfer';
 import BlockUser from './msg/profiles/block_user';
 import UnBlockUser from './msg/profiles/unblock_user';
-import CreateClient from './msg/client/create_client';
-import UpdateClient from './msg/client/update_client';
-import UpgradeClient from './msg/client/upgrade_client';
-import SubmitMisbehaviour from './msg/client/submit_misbehaviour';
-import Height from './msg/client/height';
-import Acknowledgement from './msg/channel/acknowledgement';
-import Channel from './msg/channel/channel';
-import ChannelCloseConfirm from './msg/channel/channel_close_confirm';
-import ChannelCloseInit from './msg/channel/channel_close_init';
-import ChannelOpenAck from './msg/channel/channel_open_ack';
-import ChannelOpenConfirm from './msg/channel/channel_open_confirm';
-import ChannelOpenInit from './msg/channel/channel_open_init';
-import ChannelOpenTry from './msg/channel/channel_open_try';
-import CounterpartyChannel from './msg/channel/counterparty';
-import Packet from './msg/channel/packet';
-import ReceivePacket from './msg/channel/receive_packet';
-import Timeout from './msg/channel/timeout';
-import TimeoutOnClose from './msg/channel/timeout_on_close';
-import ConnectionEnd from './msg/connection/connection_end';
-import ConnectionOpenAck from './msg/connection/connection_open_ack';
-import ConnectionOpenConfirm from './msg/connection/connection_open_confirm';
-import ConnectionOpenInit from './msg/connection/connection_open_init';
-import ConnectionOpenTry from './msg/connection/connection_open_try';
-import CounterpartyConnection from './msg/connection/counterparty';
-import Version from './msg/connection/version';
-import Transfer from './msg/transfer/transfer';
+import CreateClient from './msg/ibc/client_create_client';
+import UpdateClient from './msg/ibc/client_update_client';
+import UpgradeClient from './msg/ibc/client_upgrade_client';
+import SubmitMisbehaviour from './msg/ibc/client_submit_misbehaviour';
+import Height from './msg/ibc/client_height';
+import Acknowledgement from './msg/ibc/channel_acknowledgement';
+import Channel from './msg/ibc/channel';
+import ChannelCloseConfirm from './msg/ibc/channel_close_confirm';
+import ChannelCloseInit from './msg/ibc/channel_close_init';
+import ChannelOpenAck from './msg/ibc/channel_open_ack';
+import ChannelOpenConfirm from './msg/ibc/channel_open_confirm';
+import ChannelOpenInit from './msg/ibc/channel_open_init';
+import ChannelOpenTry from './msg/ibc/channel_open_try';
+import CounterpartyChannel from './msg/ibc/channel_counterparty';
+import Packet from './msg/ibc/channel_packet';
+import ReceivePacket from './msg/ibc/channel_receive_packet';
+import Timeout from './msg/ibc/channel_timeout';
+import TimeoutOnClose from './msg/ibc/channel_timeout_on_close';
+import ConnectionEnd from './msg/ibc/connection_end';
+import ConnectionOpenAck from './msg/ibc/connection_open_ack';
+import ConnectionOpenConfirm from './msg/ibc/connection_open_confirm';
+import ConnectionOpenInit from './msg/ibc/connection_open_init';
+import ConnectionOpenTry from './msg/ibc/connection_open_try';
+import CounterpartyConnection from './msg/ibc/connection_counterparty';
+import Version from './msg/ibc/connection_version';
+import Transfer from './msg/ibc_transfer/transfer';
 
 export {
   Overview,
