+++ src/screens/transaction_details/components/msg/ibc/channel/index.tsx
