+++ src/screens/transaction_details/components/msg/ibc/channel_close_confirm/index.tsx
