+++ src/screens/transaction_details/components/msg/ibc/channel_counterparty/index.tsx
