+++ src/screens/transaction_details/components/msg/ibc/channel_open_init/index.tsx
