+++ src/screens/transaction_details/components/msg/ibc/channel_packet/index.tsx
