+++ src/screens/transaction_details/components/msg/ibc/channel_receive_packet/index.tsx
