+++ src/screens/transaction_details/components/msg/ibc/channel_timeout_on_close/index.tsx
