+++ src/screens/transaction_details/components/msg/ibc/connection_counterparty/index.tsx
