+++ src/screens/transaction_details/components/msg/ibc/connection_open_ack/index.tsx
