+++ src/screens/transaction_details/components/msg/ibc/connection_open_confirm/index.tsx
