+++ src/screens/transaction_details/components/msg/ibc/connection_version/index.tsx
